Create a lower level container to manage the state of Article page
